web_app/requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
beanie==1.29.0
black==25.1.0
certifi==2025.4.26
click==8.1.8
croniter==6.0.0
dnspython==2.7.0
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.1
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-model==0.2.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.7.1
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
qdrant-client==1.14.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.2
yarl==1.20.0
